fix: some oom and handle table in order

+++ config/config.go
@@ -52,7 +52,7 @@ type Config struct {
 	// Databend configuration
 	DatabendDSN      string `json:"databendDSN" default:"localhost:8000"`
 	DatabendTable    string `json:"databendTable"`
-	BatchSize        int    `json:"batchSize" default:"1000"`
+	BatchSize        int64  `json:"batchSize" default:"1000"`
 	BatchMaxInterval int    `json:"batchMaxInterval" default:"3"` // for rate limit control
 
 	// related docs: https://docs.databend.com/sql/sql-commands/dml/dml-copy-into-table
